requirements.txt
torch
numpy
pandas
matplotlib
pyyaml
tqdm
scipy